backend/pyproject.toml
[project]
name = "lark-ticket"
version = "0.1.0"
description = "lark ticket's extended features."
authors = [
    { name = "D.M.M.", email = "mydevops@163.com" }
]
dependencies = [
    "gunicorn==22.0.0",
    "loguru==0.7.2",
    "httpx==0.27.0",
    "aiomysql==0.2.0",
    "uvicorn[standard]==0.30.5",
    "sqlalchemy[asyncio]==2.0.32",
    "fastapi==0.112.0",
    "orjson==3.10.6",
    "uvicorn-worker==0.2.0",
    "sentry-sdk[fastapi]==2.12.0",
    "asyncer==0.0.7",
    "pycryptodome==3.20.0",
    "lark-oapi==1.3.0",
]
requires-python = ">= 3.12"

[tool.rye]
dev-dependencies = [
    "pre-commit==3.8.0",
]

[[tool.rye.sources]]
name = "default"
url = "https://mirrors.aliyun.com/pypi/simple/"

[tool.rye.scripts]
lint = "pre-commit run --all-files"
dev = "python -m src"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.12
target-version = "py312"

[lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint]
extend-select = ["I", "UP", "D"]

[tool.ruff.lint.isort]
force-single-line = true
lines-after-imports = 2

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
